ensure assemble runs even without training section
