+++ spacy/cli/_util.py
@@ -516,18 +516,4 @@ def setup_gpu(use_gpu: int) -> None:
             msg.info("To switch to GPU 0, use the option: --gpu-id 0")
 
 
-def create_before_to_disk_callback(
-    callback: Optional[Callable[["Language"], "Language"]]
-) -> Callable[["Language"], "Language"]:
-    from ..language import Language  # noqa: F811
 
-    def before_to_disk(nlp: Language) -> Language:
-        if not callback:
-            return nlp
-        modified_nlp = callback(nlp)
-        if not isinstance(modified_nlp, Language):
-            err = Errors.E914.format(name="before_to_disk", value=type(modified_nlp))
-            raise ValueError(err)
-        return modified_nlp
-
-    return before_to_disk
